+++ CHANGELOG.md
@@ -2,4 +2,8 @@
 
 ## 0.1.0 (July 13, 2018)
 
-- combine rainbow-brackets and vscode-scheme
+- Combine rainbow-brackets and vscode-scheme
+
+### 0.1.1 (July 15, 2018)
+
+- Decolor the brackets in string object